requirements.txt
absl-py==0.13.0
antlr4-python3-runtime==4.8
anyio==3.3.1
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bert-embedding==1.0.1
bitarray==2.6.0
bleach==4.1.0
blis==0.7.8
brotlipy==0.7.0
cachetools==4.2.2
catalogue==2.0.8
certifi==2021.5.30
click==8.1.3
colorama==0.4.5
conda==4.10.3
cycler==0.10.0
cymem==2.0.6
Cython==0.29.32
dataclasses==0.6
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
fairseq==0.10.2
ftfy==6.1.1
gensim==4.2.0
gluonnlp==0.6.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
graphviz==0.8.4
grpcio==1.40.0
h5py==3.4.0
hydra-core==1.0.7
imageio==2.21.1
importlib-resources==5.9.0
info-nce-pytorch==0.1.4
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyterlab==3.1.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
langcodes==3.2.1
lxml==4.8.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.7
mxnet==1.4.0
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.8.6
nltk==3.7
notebook==6.4.4
numpy==1.21.5
oauthlib==3.1.1
omegaconf==2.0.6
opencv-python==4.6.0.66
packaging==21.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.2
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
portalocker==2.5.1
preshed==3.0.6
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==4.24.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pydantic==1.9.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
regex==2022.7.25
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
sacrebleu==2.2.0
scikit-image==0.16.2
scipy==1.9.0
Send2Trash==1.8.0
smart-open==5.2.1
sniffio==1.2.0
spacy==3.4.1
spacy-legacy==3.0.9
spacy-loggers==1.0.3
srsly==2.4.4
supervisor==4.2.2
tabulate==0.8.10
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.6.2.2
terminado==0.12.1
testpath==0.5.0
thinc==8.1.0
torch-scatter==2.0.9
torchaudio==0.9.0
tornado==6.1
traitlets==5.1.0
typer==0.4.2
typing==3.6.6
typing-extensions==3.10.0.2
tzdata==2023.3
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
zipp==3.8.1
